tests/turbohtml-tests/template_colgroup.dat

#data
<table><colgroup><template><div>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>
|         <template>
|           content
|             <div>
